recommender/requirements.txt
Django
djangorestframework
psycopg2-binary
drf-spectacular
Pillow
pydantic
typing
rich
python-decouple
tiktoken
pandas
django-cors-headers
Pillow
requests
pinecone-client
gunicorn
python-dotenv
azure-cosmos
pytest
pytest-mock
tvdb_v4_official
google-api-python-client
git+https://github.com/internetarchive/openlibrary-client.git
django-extensions
model-mommy
model-mommy
Celery
Django
django-celery-beat
django-celery-results
redis
channels
channels-redis
websockets